refactor the arrow layout
