+++ tensorflow_io/bigquery/kernels/bigquery_lib.h
@@ -43,7 +43,7 @@ limitations under the License.
 #include "tensorflow/core/lib/core/status.h"
 #include "tensorflow/core/platform/logging.h"
 #include "tensorflow/core/public/version.h"
-#include "tensorflow_io/arrow/kernels/arrow_util.h"
+#include "tensorflow_io/core/kernels/arrow/arrow_util.h"
 
 namespace tensorflow {
