+++ tensorflow_io/core/kernels/arrow/arrow_util.cc
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow_io/arrow/kernels/arrow_util.h"
+#include "tensorflow_io/core/kernels/arrow/arrow_util.h"
 
 #include "arrow/adapters/tensorflow/convert.h"
 #include "arrow/api.h"
